% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/domainAbun.R
\name{domainAbun}
\alias{domainAbun}
\title{Calculate sampling domain level index of abundance}
\usage{
domainAbun(rvcObj)
}
\arguments{
\item{rvcObj}{An RVC object (see \code{\link{rvcData}})}
}
\value{
A data.frame with the species, year, and protected
status (if applicable) as well as
\item{yt}{Index of abundance}
\item{vbar}{The variance associated with the abundance}
\item{se}{Standard error associated with the abundance}
}
\description{
Calculates domain level indices of abundance
based on the density per secondary sampling unit, and the
number of sampling units throughout the domain
}
\seealso{
\code{\link{rvcData}} \code{\link{stratAbun}} \code{\link{domain}}
}

