% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/domainLenFreq.R
\name{domainLenFreq}
\alias{domainLenFreq}
\title{Domain level length frequencies}
\usage{
domainLenFreq(rvcObj)
}
\arguments{
\item{rvcObj}{An RVC object (see \code{\link{rvcData}})}
}
\value{
A data.frame with lengths and their frequencies for
every species in every sampling domain in rvcObj
}
\description{
Domain level length frequencies
}
\seealso{
\code{\link{stratLenFreq}} \code{\link{rvcData}}
}

