% Generated by roxygen2 (4.0.2): do not edit by hand
\name{domainMaturity}
\alias{domainMaturity}
\title{Calculates domain level estimates of density or
occurrence for juveniles/adults}
\usage{
domainMaturity(matObj, calc = "d")
}
\arguments{
\item{matObj}{A MAT object, see \code{\link{maturityData}} for more details.}
}
\value{
A data.frame with the species, year, and stratum
information as well as:
\item{yi.m}{The average density/occurrence for adults}
\item{yi.j}{The average density/occurrence for juveniles}
\item{variance.m}{The variance in density/occurrence for adults}
\item{variance.j}{The variance in density/occurrence for juveniles}
\item{se.m}{The standard error of density/occurrence for adults}
\item{se.j}{The standard error of density/occurrence for juveniles}
\item{cv.m}{The coeficient of variation (as a percent) in density/occurrence for adults}
\item{n}{The total number of primary samples per sampling domain}
\item{nm}{The total number of secondary samples per sampling domain}
\item{NMTOT}{The total possible number of secondary samples per sampling domain}
}
\description{
Calculates domain level estimates of density or
occurrence for juveniles/adults
}
\seealso{
\code{\link{stratMaturity}} \code{\link{maturityData}}
}

