% Generated by roxygen2 (4.0.2): do not edit by hand
\name{domainNStar}
\alias{domainNStar}
\title{The optimal number of primary sampling units}
\usage{
domainNStar(cv, rvcObj)
}
\arguments{
\item{cv}{A single integer representing the target coefficient variance
as a percent}

\item{rvcObj}{An RVC object (see \code{\link{rvcData}})}
}
\description{
Outputs a data.frame of the optimal
number of primary sampling units per sampling domain
(n*)
}

