% Generated by roxygen2 (4.0.2): do not edit by hand
\name{domainOccurrence}
\alias{domainOccurrence}
\title{Calculate occurrence of fish species at the sampling domain
level}
\usage{
domainOccurrence(sample_data, stratum_data, species, years = "all",
  strata = "all", includes_protected = FALSE)
}
\arguments{
\item{sample_data}{A data.frame or list containing the Reef Visual Census
sample data. Must include variables named "SPECIES_CD", "YEAR",
 "STRAT", "PRIMARY_SAMPLE_UNIT", "STATION_NR", "NUM"
 (not case-sensitive).}

\item{stratum_data}{A data.frame or list containing the stratum information.
Must include variables names "YEAR", "STRAT", "NTOT"
(not case-sensitive).}

\item{species}{A character vector of scientific names or species codes,
e.g. 'Lutjanus griseus' or 'LUT GRIS' (not case-sensitive).}

\item{years}{A numeric vector of the years to select from data.
Default value 'all' selects for all years present in data.}

\item{strata}{A character vector of the strata to select from data.
Default value of 'all' selects for all strata in data.}

\item{includes_protected}{A boolean: FALSE (default) does not differentiate between
protected and unprotected areas within strata, TRUE does.}
}
\value{
A data.frame with the average occurrence, sample variance, standard error (se),
and coefficient of variance (cv), number of samples (nm), and number
of possible samples (NMTOT) per species, per year
}
\description{
Returns a data.frame with the sampling domain level occurrence (per 177m^2)
for the given species, years, and strata
}
\note{
This is a higher level version of the \code{\link{domain}}
function
}
\seealso{
\code{\link{domain}} \code{\link{stratOccurrence}} \code{\link{domainDensity}}
}

