% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getDiversityData.R
\name{getDiversityData}
\alias{getDiversityData}
\title{Retrieves Diversity data from the server}
\usage{
getDiversityData(year, region, stratum = NULL,
  server = "http://www.sefsc.noaa.gov/rvc_analysis/")
}
\arguments{
\item{year}{A numeric vector of the years to select from data.}

\item{region}{A string of the region code to select from
data. Only one region allowed at a time.}

\item{server}{A string indicating the location of the server. Can
be an IP address or URL}
}
\description{
Returns a dataframe of species richness data per secondary sampling
unit for the provided year and region
}

