% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getLhp.R
\name{getLhp}
\alias{getLhp}
\title{Returns life history parameter data}
\usage{
getLhp(species, server = "http://localhost:3000")
}
\arguments{
\item{species}{A character vector of scientific names or species codes,
e.g. 'Lutjanus griseus' or 'LUT GRIS' (not case-sensitive).}

\item{server}{A string indicating the location of the server. Can
be an IP address or URL}
}
\value{
Returns a data.frame with:
\item{SPECIES_CD}{The species code: the first three
letters of the genus name and the first four of the
species name}
\item{LC}{The minimum length at capture in centimeters.
Generally corresponds to the minimum size limit}
\item{LM}{The median length at maturity in
centimeters}
\item{WLEN_A}{The alpha coefficient of the allometric
growth equation in g ~ mm}
\item{WLEN_B}{The beta coefficient of the allometric
growth equation in g ~ mm}
}
\description{
Returns a table of life history parameters given a single
or several species and a server location
}
\note{
The allometric growth equation:
\deqn{W(g) = \alpha L(mm)^\beta }
}
\examples{
## Retrieves life history parameters for Black Grouper
bg  <- getLhp(species = "Myc Bona")
}
\seealso{
\code{\link{rvcData}} \code{\link{getStratumData}}
\code{\link{getSampleData}}
}

