% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getStat.R
\name{getStat}
\alias{getStat}
\title{Returns summary statistics}
\usage{
getStat(x, level, stat, growth_parameters = NULL, merge_protected = FALSE,
  when_present = FALSE, length_class = NULL, ...)
}
\arguments{
\item{x}{An RVC object (generated by \code{\link{rvcData}})}

\item{level}{Keyword, either "stratum" or "domain" indicating at which level the
summary statistic should be calculated}

\item{stat}{Keyword, indicating which summary statistic to calculate. Options are
"density", "occurrence", "abundance", "length_frequency", and "biomass"}

\item{growth_parameters}{A list of allometric growth parameters named a (the linear coefficient)
and b (the exponent). Only needed if stat = "biomass". Defualt value is NULL}

\item{merge_protected}{Boolean: Indicates whether protected and unprotected areas are merged together in calculating
the statistic. Default value is FALSE.}

\item{when_present}{Boolean: Indicates whether statistic is to be calculated for non-zero data, when the species
was present. NOTE: Can only be used for density and with only one species.}

\item{length_class}{Number: A breakpoint indiating the length about which to calculate the statistic.}

\item{...}{Optional parameters to pass to the select method (see \code{\link{select}})}
}
\value{
A data frame of the summary statistics
}
\description{
Returns summary statistic 'stat' at level 'level'
for RVC object 'x'
}
\seealso{
\code{\link{rvcData}} \code{\link{select}}
}

