% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getStratumData.R
\name{getStratumData}
\alias{getStratumData}
\title{Retrieves stratum data from server}
\usage{
getStratumData(year, region, stratum = NULL, protected = NULL,
  server = "http://www.sefsc.noaa.gov/rvc_analysis/")
}
\arguments{
\item{year}{A numeric vector of the years to select from data.}

\item{region}{A character vector of the region(s) from which to get sample data.
\emph{Multiple regions may be selected}.}

\item{stratum}{A four-letter code representing the stratum}

\item{protected}{An integer representing the protected status.}

\item{server}{A string indicating the location of the server. Can
be an IP address or URL}
}
\description{
Returns a data frame of stratum data pulled from the
server. Useful for writing your own function using the RVC data.
}
\examples{
## Get stratum data for protected areas in
## the Florida Keys in 2012
ntot2012  <- getStratumData(year = 2012,
region = 'fla keys', protected = TRUE);
}
\seealso{
\code{\link{rvcData}} \code{\link{getSampleData}}
}

