% Generated by roxygen2 (4.0.2): do not edit by hand
\name{maturityData}
\alias{maturityData}
\title{Creates a MAT object}
\usage{
maturityData(sample_data, stratum_data, species, lm, ...)
}
\arguments{
\item{sample_data}{A data.frame or list containing the Reef Visual Census
sample data. Must include variables named "SPECIES_CD", "YEAR",
 "STRAT", "PRIMARY_SAMPLE_UNIT", "STATION_NR", "NUM"
 (not case-sensitive).}

\item{stratum_data}{A data.frame or list containing the stratum information.
Must include variables names "YEAR", "STRAT", "NTOT"
(not case-sensitive).}

\item{species}{A single species name}

\item{lm}{Length-at-maturity for the given species}

\item{\dots}{Optional arguments that can be passed to \code{\link{rvcData}}}
}
\value{
A MAT object containing two RVC objects, one for mature individuals
and one for juveniles (see \code{\link{rvcData}} for more on RVC objects)
}
\description{
Creates a MAT object from RVC sample data and stratum data
that contains data for mature and juvenile individuals seperately
}
\seealso{
\code{\link{rvcData}}
}

