% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mStar}
\alias{mStar}
\title{Calculates the optimal number of secondary sample units per stratum}
\usage{
mStar(rvcObj)
}
\arguments{
\item{rvcObj}{An RVC object (see \code{\link{rvcData}})}
}
\description{
Calculates the optimal number of secondary sample units per
primary sample unit per stratum
}
\seealso{
\code{\link{domainNStar}} \code{\link{stratNStar}}
}

