% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/region.R
\name{region}
\alias{region}
\title{Region level stats}
\usage{
region(rvcObj, calc = "d")
}
\arguments{
\item{rvcObj}{An RVC object (see \code{\link{rvcData}})}

\item{calc}{A character: 'd' to calculate density  estimates(default),
or 'p' for occurrence estimates}
}
\description{
Density/Occurence by region
}

