\name{rvcstats-package}
\alias{rvcstats-package}
\alias{rvcstats}
\docType{package}
\title{
A statistical package for South Florida's Reef Visual Census
}
\description{
rvcstats takes assesment ready data available from the Southeast Fisheries Science Center (SEFSC) and produces density data and graphs using standardized methodologies approved by the SEFSC and its partners to help fisheries managers and scientists answer questions relevant to the state of South FLorida's reef fish community
}
\details{
\tabular{ll}{
Package: \tab rvcstats\cr
Type: \tab Package\cr
Version: \tab 0.0.1 \cr
Date: \tab 2014-09-14\cr
License: \tab GPLv3\cr
}
%ToDo: Add overview of how to use package and its most important functions
}
\author{
Harry Ganz 

Maintainer: Harry Ganz \href{mailto:hganz@rsmas.miami.edu}{hganz@rsmas.miami.edu}
}
\references{
National Park Service (2009). "A Cooperative Multi-Agency Reef Fish Monitoring Protocol for the Florida Keys Coral Reef Ecosystem". Natural Resources Report NPS/SFCN/NRR -- 2009/150 \href{http://dep.state.fl.us/coastal/programs/coral/documents/2012/FIM/04-13/FLKeys_Reef_Fish_monitoring_protocol.pdf}{http://dep.state.fl.us/coastal/programs/coral/documents/2012/FIM/04-13/FLKeys_Reef_Fish_monitoring_protocol.pdf}
}
\keyword{
package
rvc
rvcstats
}
\seealso{
% ToDo: Optional links to other man pages, e.g.
}
\examples{
  % ToDo: simple examples of the most important functions 
}
