% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{rvcstats}
\alias{rvcstats}
\alias{rvcstats-package}
\title{A statistical package for South Florida's Reef Visual Census}
\description{
Provides fisheries managers and scientists with
 summary statistics for South Florida's Reef Visual Census
 using standard statistical procedures
}
\author{
\href{mailto:harry.ganz@noaa.gov}{Harry Ganz}
}

