% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Subsets RVC object}
\usage{
select(x, species = NULL, year = NULL, region = NULL, stratum = NULL,
  protected = NULL, ...)
}
\arguments{
\item{x}{An RVC object (generated by \code{\link{rvcData}})}

\item{species}{A character vector of scientific names or species codes,
e.g. 'Lutjanus griseus' or 'LUT GRIS' (not case-sensitive).}

\item{year}{A numeric vector of the years to select from data.}

\item{region}{A character vector of the region code(s) to select from
data.}

\item{stratum}{A four-letter code representing the stratum}

\item{protected}{An integer representing the protected status.}

\item{...}{Optional parameters passed to \code{\link{rvcData}}, e.g. server}
}
\value{
Returns an RVC object with two elements:
\item{sample_data}{Contains the original sample data subsetted by
input arguments}
\item{stratum_data}{Contains original stratum data subsetted by
input arguments}
}
\description{
Produces a subset of Reef Visual Census data.
}
\seealso{
\code{\link{rvcData}} \code{\link{getStat}}
}

