% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/strat.R
\name{strat}
\alias{strat}
\title{Stratum level estimates of density/occurrence}
\usage{
strat(rvcObj, calc = "d")
}
\arguments{
\item{rvcObj}{An RVC object (see \code{\link{rvcData}})}

\item{calc}{A character: 'd' to calculate density  estimates(default),
or 'p' for occurrence estimates}
}
\value{
A data.frame containing the year, species, and stratum info as
 well as:
 \item{wh}{The weighting factor for a particular stratum}
 \item{yi}{The average density/occurrence}
 \item{mbar}{The average number of secondary sample units per stratum}
 \item{n}{The number of primary sample units per stratum}
 \item{v1}{The variance between primary sample units in density/occurrence}
 \item{v2}{The stratum-level variance between secondary sample units in density/occurrence}
 \item{s}{Sample standard deviation of density/occurrence in each stratum}
 \item{vbar}{Variance in mean density/occurrence in each stratum}
 \item{nm}{The total number of secondary sample units per stratum}
 \item{NMTOT}{The total possible number of secondary samples per stratum}
}
\description{
Outputs the stratum level estimates of
density or occurrence from an RVC object as a data.frame
}
\seealso{
\code{\link{rvcData}} \code{\link{domain}}
}

