% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stratAbun.R
\name{stratAbun}
\alias{stratAbun}
\title{Calculates stratum level indices of abundance}
\usage{
stratAbun(rvcObj)
}
\arguments{
\item{rvcObj}{An RVC object (see \code{\link{rvcData}})}
}
\value{
A data.frame with year, species, stratum, protected
status (if applicable) and:
\item{wh}{The weighting factor for a particular stratum}
 \item{yt}{The average abundance in each stratum}
 \item{vbar}{Variance in mean abundance in each stratum}
}
\description{
Calculates the stratum level index of abundance
based on the density per secondary sample unit, and the number
of possible secondary sampling units per stratum.
}
\note{
Indices of abundance produced by this function assume
homogeneity of density within strata, and should not be
used naively for assessment.
}
\seealso{
\code{\link{rvcData}} \code{\link{strat}} \code{\link{domainAbun}}
}

