% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stratLengthSummary.R
\name{stratLengthSummary}
\alias{stratLengthSummary}
\title{Summarizes Length Frequency Data for each Stratum}
\usage{
stratLengthSummary(rvcObj)
}
\arguments{
\item{rvcObj}{An RVC object (see \code{\link{rvcData}})}
}
\value{
A data.frame with SPECIES_CD, YEAR, STRAT, PROT (if applicable), and:
\item{min}{A numeric vector of the minimum length}
\item{first.quartile}{A numeric vector of the first quartile of length}
\item{mean}{A numeric vector of the mean length}
\item{median}{A numeric vector of the median length}
\item{third.quartile}{A numeric vector of the third quartile of length}
\item{max}{A numeric vector of the maximum length}
}
\description{
Produces a data.frame with summary statistics
of length split by species, year, and protected status
(if applicable)
}
\seealso{
\code{\link{stratLenFreq}}
}

