% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stratNStar}
\alias{stratNStar}
\title{The optimal number of primary sampling units per stratum}
\usage{
stratNStar(cv, rvcObj)
}
\arguments{
\item{cv}{A single integer representing the target coefficient variance
as a percent}

\item{rvcObj}{An RVC object (see \code{\link{rvcData}})}
}
\value{
A data.frame including the species code, year,
stratum, and nstar for the target cv
}
\description{
Outputs a data.frame the optimal number of sampling units
per stratum for a given coeeficient of variance
}

