% Generated by roxygen2 (4.0.2): do not edit by hand
\name{stratOccurrence}
\alias{stratOccurrence}
\title{Calculate occurrence of fish at the stratum level}
\usage{
stratOccurrence(sample_data, stratum_data, species, years = "all",
  strata = "all", includes_protected = FALSE)
}
\arguments{
\item{sample_data}{A data.frame or list containing the Reef Visual Census
sample data. Must include variables named "SPECIES_CD", "YEAR",
 "STRAT", "PRIMARY_SAMPLE_UNIT", "STATION_NR", "NUM"
 (not case-sensitive).}

\item{stratum_data}{A data.frame or list containing the stratum information.
Must include variables names "YEAR", "STRAT", "NTOT"
(not case-sensitive).}

\item{species}{A character vector of scientific names or species codes,
e.g. 'Lutjanus griseus' or 'LUT GRIS' (not case-sensitive).}

\item{years}{A numeric vector of the years to select from data.
Default value 'all' selects for all years present in data.}

\item{strata}{A character vector of the strata to select from data.
Default value of 'all' selects for all strata in data.}

\item{includes_protected}{A boolean: FALSE (default) does not differentiate between
protected and unprotected areas within strata, TRUE does.}
}
\value{
A data.frame with the average occurrence, sample variance,
number of primary sample units (n), number of secondary sample units (nm),
number of possible primary sample units (NTOT), and number of possible
secondary sample units (NMTOT)
}
\description{
Returns a data.frame with the stratum level average occurrence (per 176m^2),
for the given species per species per year
}
\note{
A higher level version of the \code{\link{strat}} function
}
\seealso{
\code{\link{domainOccurrence}} \code{\link{strat}} \code{\link{stratDensity}}
}

