% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/domain.R
\name{domain}
\alias{domain}
\title{Domain level estimates of density/occurrence}
\usage{
domain(rvcObj, stat = "density")
}
\arguments{
\item{rvcObj}{An RVC object (see \code{\link{rvcData}})}

\item{stat}{A String: "density" to calculate density  estimates(default),
or "occurrence" for occurrence estimates}
}
\value{
A data.frame with the species, year, and stratum
information as well as:
\item{yi}{The average density/occurrence}
\item{variance}{The variance in density/occurrence}
\item{se}{The standard error of density/occurrence}
\item{cv}{The coeficient of variation (as a percent) in density/occurrence}
\item{n}{The total number of primary samples per sampling domain}
\item{nm}{The total number of secondary samples per sampling domain}
\item{NMTOT}{The total possible number of secondary samples per sampling domain}
}
\description{
Outputs the domain level estimates of density or
occurrence given an RVC object
}
\seealso{
\code{\link{rvcData}} \code{\link{strat}}
}

