% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSampleData.R
\name{getSampleData}
\alias{getSampleData}
\title{Retrieves RVC sample data from server}
\usage{
getSampleData(species, year, region, stratum = NULL, protected = NULL,
  when_present = NULL, server)
}
\arguments{
\item{species}{A character vector of scientific names or species codes,
e.g. 'Lutjanus griseus' or 'LUT GRIS' (not case-sensitive).}

\item{year}{A numeric vector of the years to select from data.}

\item{region}{A character vector of the region(s) from which to get sample data.
\emph{Multiple regions may be selected}.}

\item{stratum}{A four-letter code representing the stratum}

\item{protected}{An integer representing the protected status.}

\item{when_present}{A boolean indicating whether only samples where
individuals were seen should be selected. \cr
\emph{Multiple species may be selected}}

\item{server}{A string indicating the location of the server. Can
be an IP address or URL}
}
\description{
Returns a data frame of the sample data pulled from the server. Useful for
analyzing the analysis ready data using custom functions not included
in this package.
}
\seealso{
\code{\link{rvcData}} \code{\link{getStratumData}}
}

