% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getStat.R
\name{getStat}
\alias{getStat}
\title{Summary Statistics for RVC data}
\usage{
getStat(x, level, stat, growth_parameters = NULL, merge_protected = TRUE,
  when_present = FALSE, length_class = NULL, bin_width = NULL, ...)
}
\arguments{
\item{x}{An RVC object (generated by \code{\link{rvcData}})}

\item{level}{Keyword: either "stratum" or "domain" indicating at which level the
summary statistic should be calculated}

\item{stat}{Keyword: indicating which summary statistic to calculate. Options are
"density", "occurrence", "abundance", "length_frequency", and "biomass"}

\item{growth_parameters}{\cr
\describe{
   \item{If x contains a single species:}{
   A list of allometric growth parameters named 'a' (the linear coefficient)
   and 'b' (the exponent).
   If no growth parameters are provided (NULL),
   function will attempt to retrieve them from RVC object.}
   \item{If x contains multiple species:}{
   A list of lists, with each key a species code with underscores instead of spaces and
   each value a list of growth parameters in the same format as for a single species. \cr
   e.g. \code{growth_parameters = list(EPI_MORI = list(a=1.2e-5, b=2.88), MYC_BONA = list(a=1.88e-5, b=3))}. \cr
   If a species is value is NULL, function will attempt to retrieve growth_parameters
   from RVC object. If not NULL, all species present must be listed.}
 }
 \strong{NOTE:} Argument format is dependent on number of species in RVC object x. If
 x contains multiple species, but only one species is selected via the optional ... parameter
 the multispecies form must still be used. Also, multispecies form cannot be used for
 single species.  \cr
 \strong{NOTE:} Units for growth parameters are in mm not cm}

\item{merge_protected}{Boolean: Indicates whether protected and unprotected areas are merged together in calculating
the statistic. Default value is TRUE.}

\item{when_present}{Boolean: Indicates whether statistic is to be calculated for non-zero data, when the species
was present.}

\item{length_class}{\cr
\describe{
 \item{If x contains a single species:}{
   Number or Keyword: indicating a break point between two length classes, such as the breakpoint between
   immature and mature individuals or non-exploitable and exploitable individuals. The recognized
   keywords are "LM" for median length-at-maturity and "LC" for minimum length-at-capture. If keywords
   are provided, getStat will attempt to retrieve the breakpoint values from the RVC object, otherwise it will
   use the provided breakpoint. }
   \item{if x contains multiple species}{
   A list, with each key a species code with underscores instead of spaces and
   each value a number or keyword in the same format as for a single species. \cr
   e.g. \code{length_class = list(EPI_MORI = "LC", MYC_BONA = 62)}. \cr
   If a species is not listed or value is NULL, function will calculate statistic for
   all length classes combined.
   }
 }
   \strong{NOTE:} Break is non-inclusive for the lower interval and inclusive for the upper (i.e. lower > break >= upper).}

\item{bin_width}{A number indicating the width of each length bin is stat="length_frequency".
If NULL (default) lengths will not be binned. \cr
\strong{NOTE:} In output, bins will be labelled by their midpoints.}

\item{...}{Optional parameters to pass to the select method (see \code{\link{select}})}
}
\value{
Returns: a data frame of the summary statistics
}
\description{
Returns a data.frame of summary statistics given an RVC object
}
\examples{
x  <- rvcData(species = 'EPI MORI', year = 2012, region = 'FLA KEYS')
## Calculate Domain-Level Density
getStat(x, level = "domain", stat = "density")
## Calculate Stratum Level biomass
## with protected and unprotected areas listed separately
getStat(x, level = "stratum", stat = "biomass",
 growth_parameters = list(a = 0.061, b = 2.2),
 merge_protected = FALSE)
## Calculate domain level density for individuals above and below
## 40cm
getStat(x, level = "domain", stat = "density", length_class = 40)
}
\seealso{
\code{\link{rvcData}} \code{\link{select}}
}

