% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getStat.R
\name{getStat}
\alias{getStat}
\title{Returns summary statistics}
\usage{
getStat(x = NULL, level, stat, growth_parameters = NULL, ...)
}
\arguments{
\item{x}{An RVC object (generated by \code{\link{rvcData}})}

\item{level}{Keyword, either "stratum" or "domain" indicating at which level the
summary statistic should be calculated}

\item{stat}{Keyword, indicating which summary statistic to calculate. Options are
"density", "occurrence", "abundance", "length_frequency", and "biomass"}

\item{growth_parameters}{Optional: a list of allometric growth parameters named a (the linear coefficient)
and b (the exponent)}

\item{...}{Optional parameters to pass to the select method (see \code{\link{select}})}
}
\value{
A data frame of the summary statistics
}
\description{
Returns summary statistic 'stat' at level 'level'
for RVC object x
}
\seealso{
\code{\link{rvcData}} \code{\link{select}}
}

