% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rvcData.R
\name{rvcData}
\alias{rvcData}
\title{An RVC data object}
\usage{
rvcData(species, year, region, server = "http://localhost:3000")
}
\arguments{
\item{species}{A character vector of scientific names or species codes,
e.g. 'Lutjanus griseus' or 'LUT GRIS' (not case-sensitive).}

\item{year}{A numeric vector of the years to select from data.
Default value NULL selects for all years present in data.}

\item{region}{A character vector of the region code(s) to select from
data.}

\item{...}{Optional parameters passed to rvcData, e.g. server}
}
\value{
Returns an RVC object with two elements:
\item{sample_data}{Contains the original sample data subsetted by
input arguments}
\item{stratum_data}{Contains original stratum data subsetted by
input arguments}
}
\description{
Creates an RVC data object from Reef Visual Census
sample and stratum data
}

