% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rvcData.R
\name{rvcData}
\alias{rvcData}
\title{Get RVC data from server}
\usage{
rvcData(species, year, region, server = "http://localhost:3000")
}
\arguments{
\item{species}{A character vector of scientific names or species codes,
e.g. 'Lutjanus griseus' or 'LUT GRIS' (not case-sensitive).}

\item{year}{A numeric vector of the years to select from data.}

\item{region}{A character vector of the region code(s) to select from
data.}
}
\value{
Returns an RVC object with two elements:
\item{sample_data}{Contains the original sample data subsetted by
input arguments}
\item{stratum_data}{Contains original stratum data subsetted by
input arguments}
}
\description{
Creates an RVC data object from Reef Visual Census
sample and stratum data
}
\examples{
## Names not case-sensitive
rvcData(species = 'Epinephelus morio', year = 2012, region = 'FLA KEYS')
## Can pass multiple species and years
rvcData(c('LUT GRIS', 'LUT ANAL'), year = c(2010, 2012), region = 'FLA KEYS',
server = '127.0.0.1:3000')
}
\seealso{
\code{\link{select}} \code{\link{getStat}}
}

