% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rvcData}
\alias{rvcData}
\title{An RVC data object}
\usage{
rvcData(sample_data, stratum_data, species, length_class = "all",
  when_present = FALSE, years = "all", strata = "all",
  includes_protected = FALSE)
}
\arguments{
\item{sample_data}{A data.frame or list containing the Reef Visual Census
sample data. Must include variables named "SPECIES_CD", "YEAR",
 "STRAT", "PRIMARY_SAMPLE_UNIT", "STATION_NR", "NUM"
 (not case-sensitive).}

\item{stratum_data}{A data.frame or list containing the stratum information.
Must include variables names "YEAR", "STRAT", "NTOT"
(not case-sensitive).}

\item{species}{A character vector of scientific names or species codes,
e.g. 'Lutjanus griseus' or 'LUT GRIS' (not case-sensitive).}

\item{length_class}{A numeric vector of length two, with the first value the minimum length and
the second the maximum length (inclusive) to include in RVC object.
Default value of 'all' selects for all lengths.}

\item{when_present}{A boolean: FALSE (default) to include stations where species
is present and absent, and TRUE to select only stations
where species is present.}

\item{years}{A numeric vector of the years to select from data.
Default value 'all' selects for all years present in data.}

\item{strata}{A character vector of the strata to select from data.
Default value of 'all' selects for all strata in data.}

\item{includes_protected}{A boolean: FALSE (default) does not differentiate between
protected and unprotected areas within strata, TRUE does.}
}
\value{
Returns an RVC object with two elements:
\item{sample_data}{Contains the original sample data subsetted by
input arguments}
\item{stratum_data}{Contains original stratum data subsetted by
input arguments}
}
\description{
Creates an RVC data object from Reef Visual Census
sample and stratum data
}

