% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rvcData.R
\name{rvcData}
\alias{rvcData}
\title{Get RVC data from server}
\usage{
rvcData(species, year, region, server = "http://localhost:3000")
}
\arguments{
\item{species}{A character vector of scientific names or species codes,
e.g. 'Lutjanus griseus' or 'LUT GRIS' (not case-sensitive).}

\item{year}{A numeric vector of the years to select from data.}

\item{region}{A string of the region code to select from
data. Only one region allowed at a time.}

\item{server}{A string indicating the location of the server. Can
be an IP address or URL}
}
\value{
Returns an RVC object with three elements:
\item{sample_data}{A data frame containing the analysis ready sample data}
\item{stratum_data}{A data frame containing data about the stratum, including the number of
possible primary sample units per stratum (NTOT)}
\item{lhp_data}{A data frame containing life history parameter data, including
minimum length-at-capture (LC), median length-at-maturity (LM), and the allometric growth
parameters (WLEN_A, WLEN_B).
 NULL if no life history data available for selected species}
}
\description{
Creates an RVC data object from Reef Visual Census
data located on a remote server
}
\note{
Best practice it to pull off the maximum amount of data you
plan on working with first using \code{\link{rvcData}},
and then to use the \code{\link{select}}
function to break it into smaller manageable bits. This reduces the
time spent downloading it from the server.
}
\examples{
## Names not case-sensitive
rvcData(species = 'Epinephelus morio', year = 2012, region = 'FLA KEYS')
## Can pass multiple species and years
rvcData(c('LUT GRIS', 'LUT ANAL'), year = c(2010, 2012), region = 'FLA KEYS',
server = '127.0.0.1:3000')
}
\seealso{
\code{\link{select}} \code{\link{getStat}}
}

