\name{rvcstats-package}
\alias{rvcstats-package}
\alias{rvcstats}
\docType{package}
\title{Reef Visual Census Statistical Package
}
\description{
An R package with functions designed to output tables, graphs, and time series of RVC data using standardized statistical procedures approved by the Southeast Fisheries Science Center. 
}
\details{
\tabular{ll}{
Package: \tab rvcstats\cr
Type: \tab Package\cr
Version: \tab 0.2.1\cr
Date: \tab 2014-09-14\cr
License: \tab GPLv3\cr
}
%ToDo: Offer advice on how to get started with this package

}
\author{
Harry Ganz

Maintainer: Harry Ganz \href{harry.ganz@noaa.gov}{mailto:harry.ganz@noaa.gov}

}
\references{
National Park Service (2009). "A cooperative multi-agency reef fish monitoring protocol for the florida keys ecosystem". Natural Resources Report: NPS/SFCN/NRR -- 2009/150. \href{http://www.dep.state.fl.us/coastal/programs/coral/documents/2012/FIM/04-13/FLKeys_Reef_Fish_monitoring_protocol.pdf}{http://www.dep.state.fl.us/coastal/programs/coral/documents/2012/FIM/04-13/FLKeys_Reef_Fish_monitoring_protocol.pdf}
}

\keyword{ 
rvc
rvcstats
}
\seealso{
\strong{low-level functions}:
\code{\link{rvcData}}
\code{\link{stratData}}
\code{\link{strat}}
\code{\link{domain}}
\code{\link{lenFreq}}
\code{\link{nStar}}

\strong{high-level functions}:
\code{\link{stratDensity}}
\code{\link{stratOccurrence}}
\code{\link{domainDensity}}
\code{\link{domainOccurrence}}

\strong{data-sets}:
\code{\link{rvc1}}
\code{\link{rvc2}}
\code{\link{strat1}}
\code{\link{strat2}}
}
\examples{
%ToDo examples of test functions
}


