% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Subsets RVC object}
\usage{
select(x, species = NULL, year = NULL, region = NULL, stratum = NULL,
  protected = NULL, ...)
}
\arguments{
\item{x}{An RVC object (generated by \code{\link{rvcData}})}

\item{species}{A character vector of scientific names or species codes,
e.g. 'Lutjanus griseus' or 'LUT GRIS' (not case-sensitive).}

\item{year}{A numeric vector of the years to select from data.}

\item{region}{A string of the region code to select from
data. Only one region allowed at a time.}

\item{stratum}{A four-letter code representing the stratum}

\item{protected}{An integer representing the protected status.}

\item{...}{Optional parameters passed to \code{\link{rvcData}}, e.g. server}
}
\value{
Returns an RVC object with two elements:
\item{sample_data}{Contains the original sample data subsetted by
input arguments}
\item{stratum_data}{Contains original stratum data subsetted by
input arguments}
}
\description{
Produces a subset of Reef Visual Census data.
}
\note{
\code{\link{select}} is best used to subset large amounts of data pulled off the
server in \code{\link{rvcData}} into smaller manageable chunks. See
documentation for \code{\link{rvcData}} for more details
}
\examples{
x  <- rvcData(species = c("EPI MORI", "MYC BONA"),
year = 2012, region = "FLA KEYS")
## Select 1 species from a group
select(x, species = "EPI MORI")
## Select only samples from protected areas
select(x, protected = 1)
## Select a subset of strata
select(x, stratum = c("FSLR","FMLR","FDLR"))
}
\seealso{
\code{\link{rvcData}} \code{\link{getStat}}
}

