% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Subsets RVC object}
\usage{
select(x = NULL, species = NULL, year = NULL, region = NULL,
  stratum = NULL, protected = NULL, when_present = FALSE,
  merge_protected = FALSE, length_class = NULL, ...)
}
\arguments{
\item{x}{An RVC object (generated by \code{\link{rvcData}})}

\item{species}{A character vector of scientific names or species codes,
e.g. 'Lutjanus griseus' or 'LUT GRIS' (not case-sensitive).}

\item{year}{A numeric vector of the years to select from data.
Default value NULL selects for all years present in data.}

\item{region}{A character vector of the region code(s) to select from
data.}

\item{stratum}{A four-letter code representing the stratum}

\item{protected}{An integer representing the protected status.}

\item{when_present}{A boolean. : If true selects only data where species was observed.
Throws exception if more than one species selected. Defualt
value is false}

\item{merge_protected}{A boolean. If TRUE merges protected and unprotected
strata together. Default value is false}

\item{length_class}{Can be NULL, keyword or number. Must be NULL if more than one species selected, or else throws exception
\itemize{
\item If NULL does not modify sample data
\item If keyword "maturity" or "exploitable", retrieves median length at maturity or capture (respectively)
for the provided species from the server, and uses that as a breakpoint.
If not found, throws exception
\item If number, the provided number is used as the breakpoint.
}}

\item{...}{Optional parameters passed to rvcData, e.g. server}
}
\value{
Returns an RVC object with two elements:
\item{sample_data}{Contains the original sample data subsetted by
input arguments}
\item{stratum_data}{Contains original stratum data subsetted by
input arguments}
}
\description{
Produces a subset of Reef Visual Census data.
}

