% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stratLenFreq.R
\name{stratLenFreq}
\alias{stratLenFreq}
\title{Stratum level length frequencies}
\usage{
stratLenFreq(rvcObj)
}
\arguments{
\item{rvcObj}{An RVC object (see \code{\link{rvcData}})}
}
\value{
A data.frame with lengths and their frequencies for
every species in every stratum in rvcObj
}
\description{
Stratum level length frequencies
}
\seealso{
\code{\link{domainLenFreq}} \code{\link{rvcData}}
}

